Fixes https://jira.duraspace.org/browse/DURACLOUD-895: Synctool fails when the max file size is greater than max java integer value

+++ common-rest/src/main/java/org/duracloud/common/rest/RestUtil.java
@@ -25,7 +25,7 @@ public interface RestUtil {
     public class RequestContent {
         protected InputStream contentStream = null;
         protected String mimeType = null;
-        protected int size = 0;
+        protected long size = 0;
 
         /**
          * @return the contentStream
@@ -44,7 +44,7 @@ public interface RestUtil {
         /**
          * @return the size
          */
-        public int getSize() {
+        public long getSize() {
             return size;
         }
     }
